refactored some classes
