+++ src/main/java/at/fhv/backend/messageModels/PlayerJoinMessage.java
@@ -1,5 +1,6 @@
-package at.fhv.backend.model;
+package at.fhv.backend.messageModels;
 
+import at.fhv.backend.model.Position;
 import lombok.Getter;
 import lombok.Setter;
