+++ src/main/java/at/fhv/backend/utils/MapLoader.java
@@ -1,4 +1,4 @@
-package at.fhv.backend.model;
+package at.fhv.backend.utils;
 
 import java.io.File;
 import java.io.FileNotFoundException;
